requirements.txt
absl-py          ==         0.12.0
asgiref           ==       3.3.1
astor            ==        0.8.1
blis             ==        0.7.4
boto3            ==        1.17.73
botocore         ==        1.20.73
CacheControl     ==        0.12.6
cached-property  ==        1.5.2
cachetools       ==        4.2.2
catalogue        ==        2.0.1
certifi          ==        2020.12.5
cffi             ==        1.14.5
chardet          ==        4.0.0
click            ==        7.1.2
cymem            ==        2.0.5
Cython           ==        0.29.21
DateTime         ==        4.3
Django           ==        3.1.7
djangorestframework==      3.12.2
filelock           ==      3.0.12
firebase-admin     ==      5.0.0
gast               ==      0.2.2
gensim             ==      3.6.0
google-api-core    ==      1.26.3
google-api-python-client ==2.3.0
google-auth            ==  1.30.0
google-auth-httplib2 ==    0.1.0
google-auth-oauthlib ==    0.4.4
google-cloud-core    ==    1.6.0
google-cloud-firestore ==  2.1.0
google-cloud-storage  ==   1.38.0
google-crc32c         ==   1.1.2
google-pasta         ==    0.2.0
google-resumable-media ==  1.2.0
googleapis-common-protos== 1.53.0
grpcio                ==   1.37.1
h5py                  ==   2.10.0
httplib2              ==   0.19.1
huggingface-hub       ==   0.0.8
idna                  ==   2.10
importlib-metadata    ==   4.0.1
Jinja2                ==   2.11.3
jmespath              ==   0.10.0
joblib                ==   1.0.1
Keras                 ==   2.2.5
Keras-Applications    ==   1.0.8
Keras-Preprocessing   ==   1.1.2
lxml                  ==   4.6.3
Markdown              ==   3.3.4
MarkupSafe            ==   1.1.1
msgpack               ==   1.0.2
multitasking          ==   0.0.9
murmurhash            ==   1.0.5
mysql                 ==   0.0.2
mysql-connector-python ==  8.0.23
mysqlclient            ==  2.0.3
nltk                   ==  3.5
numpy                  ==  1.19.5
oauthlib              ==   3.1.0
opt-einsum            ==   3.3.0
packaging             ==   20.9
pandas                ==   0.25.3
pathy                 ==   0.4.0
preprocessor          ==   1.1.3
preshed               ==   3.0.5
proto-plus            ==   1.18.1
protobuf              ==   3.17.0
pyasn1                ==   0.4.8
pyasn1-modules        ==   0.2.8
pycparser             ==   2.20
pydantic              ==   1.7.3
PyMySQL               ==   1.0.2
pyparsing             ==   2.4.7
PySocks               ==   1.7.1
python-dateutil       ==   2.8.1
python-decouple       ==   3.4
pytz                  ==   2021.1
PyYAML                ==   5.4.1
regex                 ==   2020.11.13
requests              ==   2.25.1
requests-oauthlib     ==   1.3.0
rsa                   ==   4.7.2
s3transfer            ==   0.4.2
sacremoses            ==   0.0.45
scikit-learn          ==   0.24.1
scipy                 ==   1.5.4
sentencepiece         ==   0.1.95
setuptools            ==   41.2.0
six                   ==   1.15.0
sklearn               ==   0.0
smart-open            ==   3.0.0
spacy                 ==   3.0.3
spacy-legacy          ==   3.0.1
SQLAlchemy            ==   1.3.23
sqlparse              ==   0.4.1
srsly                 ==   2.4.0
tensorboard           ==   2.0.2
tensorflow            ==   2.0.0
tensorflow-estimator  ==   2.0.1
termcolor             ==   1.1.0
textblob              ==   0.15.3
thinc                 ==   8.0.1
threadpoolctl         ==   2.1.0
tokenizers            ==   0.8.1rc2
tqdm                  ==   4.58.0
transformers          ==   3.1.0
tweepy                ==   3.10.0
tweet-preprocessor    ==   0.6.0
Twitter-Sentimet-65   ==   0.0.1
typer                 ==   0.3.2
typing-extensions     ==   3.10.0.0
uritemplate           ==   3.0.1
urllib3               ==   1.26.3
wasabi                ==   0.8.2
Werkzeug              ==   2.0.0
wheel                 ==   0.36.2
wrapt                  ==  1.12.1
yfinance               ==  0.1.59
zipp                   ==  3.4.1
zope.interface         ==  5.4.0